requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
python-telegram-bot==21.1.1